Fix a bug where we're handed plausible garbage as input
If I type in "Fake, XY" this gets by our simple query validation, but
straigt up fails to geocode. So, we'll raise when that happens.

If I type in "Fake, NY" this also gets by the simple query validation,
but it does get geocoded at the state level. However, there's no ZIP
code. Since the system is so ZIP code oriented, we consider this case
a failure, too.

This needs some tests.